fix: Remove to barebones importable
Move stream work out of brain.js
